Backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==5.0.1
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0
marshmallow==3.20.1
pymysql==1.1.0 